Move the `jruby` option to Gem and GemPush tasks, and clarify why they are not recommended to configure (Fix #32 and #33)
